requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-magic-bin
django-filter